now move VpcRouter and VpcRouterKind into views (compiles, tests pass)
