+++ nexus/src/external_api/http_entrypoints.rs
@@ -13,7 +13,8 @@ use crate::ServerContext;
 use super::{
     console_api, params,
     views::{
-        Organization, Project, Rack, Role, Sled, Snapshot, User, Vpc, VpcSubnet,
+        Organization, Project, Rack, Role, Sled, Snapshot, User, Vpc,
+        VpcRouter, VpcSubnet,
     },
 };
 use crate::context::OpContext;
@@ -59,7 +60,6 @@ use omicron_common::api::external::RouterRouteUpdateParams;
 use omicron_common::api::external::Saga;
 use omicron_common::api::external::VpcFirewallRuleUpdateParams;
 use omicron_common::api::external::VpcFirewallRules;
-use omicron_common::api::external::VpcRouter;
 use ref_cast::RefCast;
 use schemars::JsonSchema;
 use serde::Deserialize;
