+++ nexus/test-utils/src/resource_helpers.rs
@@ -17,10 +17,11 @@ use omicron_common::api::external::Disk;
 use omicron_common::api::external::IdentityMetadataCreateParams;
 use omicron_common::api::external::Instance;
 use omicron_common::api::external::InstanceCpuCount;
-use omicron_common::api::external::VpcRouter;
 use omicron_nexus::crucible_agent_client::types::State as RegionState;
 use omicron_nexus::external_api::params;
-use omicron_nexus::external_api::views::{Organization, Project, Vpc};
+use omicron_nexus::external_api::views::{
+    Organization, Project, Vpc, VpcRouter,
+};
 use omicron_sled_agent::sim::SledAgent;
 use std::sync::Arc;
 use uuid::Uuid;
